pyproject.toml
[project]
name = "agenticai-quantitativeanalysissystem"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-storage-blob>=12.28.0",
    "crewai>=1.10.1",
    "crewai-tools>=1.10.1",
    "fastapi>=0.135.1",
    "firecrawl-py>=4.19.0",
    "ipykernel>=7.2.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.11.10",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.48",
    "streamlit>=1.55.0",
    "uvicorn>=0.42.0",
    "yfinance>=1.2.0",
]
